Move user login routes to dedicated folder and update server.js
- Moved user authentication routes to 'routes/users.js'.

- Updated 'server.js' to use the new routes file.

- Refactored route imports and middleware setup for clarity.

This reorganization improves code modularity and maintainability by separating route handling from the main server file. It also prepares the codebase for future route additions and easier management.
